Refactor team_collection methods into team_collection_methods as a module with all tests passing
